Variable price range extrema
Used the new findMinMaxPrices() method of the RoomRepository to make the price range of the search form entirely dependent on those values so that the range extrema are set dynamically knowing the database actual information, instead of being set to arbitrary constants. By default, when no search is requested, the price range uses those extrema.

Signed-off-by: Paul Mabileau <paulmabileau@hotmail.fr>