Put submit button on the right
It was on the left because I thought if the user pressed the search submit button by mistake when manipulating the date pickers, it could be frustrating. But as I intend to add another button that would be more frustrating to misclick, I think putting this one on the left is better.

Signed-off-by: Paul Mabileau <paulmabileau@hotmail.fr>