Fixed reset button
Previously, the reset button did not put the price double slider on default positions, now it does. In order to do that, it does not use the noUiSlider.reset() method which puts handles to their start values that we control when receiving the response from the POST request, it properly sets them to the minimal and maximal values available. It also updates the hidden inputs values.

Signed-off-by: Paul Mabileau <paulmabileau@hotmail.fr>