Client or owner registration
Added the ability to register an account as a Client or (inclusive) an Owner using checkbox inputs. The form result is treated correctly and saved to the database. The Owner class had to be changed a bit in order to let the registration function well, even if no country code is given.

Signed-off-by: Paul Mabileau <paulmabileau@hotmail.fr>